Better handling of nodes with unknown tokens
Instead of managing unknown tokens in pairs, we simply do not add nodes
to pairs if they do not contain known tokens. Such nodes are now
recorded in a separate set (unusable_nodes) which we can analyze. If
there is a large number of unusable nodes, the doc2vec model needs to be
re-created with a domain specific corpus.